backend/requirements.txt
# Phase 1 — ML core
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.2.0
statsmodels>=0.14.0
joblib>=1.3.0
optuna>=3.5.0
pyyaml>=6.0.0

# Notebooks / EDA
matplotlib>=3.7.0
seaborn>=0.13.0
jupyter>=1.0.0

# Phase 2 — API (install when building backend)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0